Remove the feature that branch names can be abbreviated in the output.
It turned out to be too easy to use parentheses in the regexp for
other purposes, inadvertently causing only part of the reference name
to be output and making the results very confusing.  The abbreviation
feature might be reintroduced later, but in a more explicit form that
is harder to trigger accidentally.