Fall back to python if there is no python2, like e.g. on Windows
As msysgit does not ship with Python, a natural choice is to install
vanilla Python 2 from www.python.org on Windows. That installation does
not come with a "python2" executable, so allow to fall back to "python".
Also, on Windows the vanilla Python installation will usually not be at
/usr/bin, so search the PATH instead.

Uses the method described at:
http://softwareswirl.blogspot.com/2011/06/starting-python-script-intelligently.html