find_merge(): compute a commit graph then work from that
Instead of using individual Git commands to learn about the commit
graph, compute the whole `--ancestry-path` graph of commits once then
compute what we need from that.

This approach is noticeably slower for commits that are directly on the
branch, but a bit faster for commits that were merged in. On average it
is probably a bit slower, but not prohibitively so.

Signed-off-by: Michael Haggerty <mhagger@alum.mit.edu>